requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
certifi==2024.6.2
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
python-json-logger==2.0.7
setuptools==70.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
yarl==1.9.4

